// src/types/person.ts
export interface PersonFormValues {
  lastName: string;
  firstName: string;
  middleName?: string;
  isMiddleName: boolean;
  isInp: boolean;
  inp?: string;
  birthDate: string;
  gender: string;
  country: string;
  birthPlace: string;
  contactType: string;
  secretWord: string;
  phone: string;
  email: string;
  documentType: string;
  documentNumber: string;
  issuedBy: string;
  issuedAt?: string;
  expiresAt?: string;
  UNZR: string;
}